0.9f : corrections trouvées pendant le test complet dans le navigateur
- Sécurité : pages du jeu et de l'admin inaccessibles sans connexion, mipattack.php supprimé,
  triches de flotte (vitesse, durées, limite d'expéditions) recalculées côté serveur,
  annonce et phalange protégées par jeton, liens xnova.fr retirés, mail d'inscription sans mot de passe
- « Jeu fermé » réparé ; connexion : erreur affichée sur la page de connexion, message unique
- Admin : ajout de flotte réparé et au menu, redirections, protection des débutants réglable
- Combat : création de lune ($galenemyrow), rapport de combat affiché dans la page du jeu
- Recyclage, galaxie, statistiques, vue générale (colonies, horloge), arbre des technologies
- Mise à jour 0.9f : températures des lunes, anciens liens de rapport, jeu rouvert, adresses xnova.fr vidées
- Avertissements PHP réduits (paramètres absents, prix, pretty_time), caractères perdus retapés

+++ admin/unbanned.php
@@ -25,12 +25,12 @@ include($xnova_root_path . 'common.' . $phpEx);
 		$parse['dpath'] = $dpath;
 		$parse = $lang;
 
-		$mode = $_GET['mode'];
+		$mode = ($_GET['mode'] ?? null);
 
 		if ($mode != 'change') {
 			$parse['Name'] = "Nom du joueur";
-		} elseif ($mode == 'change') {
-			$nam = SqlEscape($_POST['nam']);
+		} elseif ($mode == 'change' && $_SERVER['REQUEST_METHOD'] == 'POST') {
+			$nam = SqlEscape(($_POST['nam'] ?? null));
 			doquery("DELETE FROM {{table}} WHERE who2='{$nam}'", 'banned');
 			doquery("UPDATE {{table}} SET bana=0, banaday=0 WHERE username='{$nam}'", "users");
 			message("Le joueur {$nam} a bien &eacute;t&eacute; d&eacute;banni!", 'Information');
